+++ adapters/unsupported/lpawin.pl
@@ -1,7 +1,7 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %
 %  Adapter file for LPA WinProlog 5.00
-%  Last updated on August 5, 2020
+%  Last updated on September 1, 2020
 %
 %  This file is part of Logtalk <https://logtalk.org/>
 %  Copyright 1998-2020 Paulo Moura <pmoura@logtalk.org>
@@ -640,8 +640,8 @@ call(F, A1, A2, A3, A4, A5, A6) :-
 
 % '$lgt_read_term'(@stream, -term, +list, -position, -list)
 
-'$lgt_read_term'(Stream, Term, _Options, '-'(-1, -1), Variables) :-
-	eread(Term, Variables) <~ Stream.
+'$lgt_read_term'(Stream, Term, [variable_names(VariablesNames)| Options], '-'(-1, -1)) :-
+	eread(Term, VariablesNames) <~ Stream.